Working version of user login and logout
